Version 2 of ban list config
